update: added organization filters for complaints, leave req, attendance and refactored code
