+++ components/staff/attendance/AttendanceHeader.tsx
@@ -8,6 +8,8 @@ type AttendanceHeaderProps = {
   setSearchTerm: (term: string) => void
   onExport: () => void
   isExporting: boolean
+  selectedCategory: string
+  handleCategoryChange: (category: string) => void
 }
 
 export function AttendanceHeader({
@@ -16,8 +18,17 @@ export function AttendanceHeader({
   searchTerm,
   setSearchTerm,
   onExport,
-  isExporting
+  isExporting,
+  selectedCategory,
+  handleCategoryChange
 }: AttendanceHeaderProps) {
+
+  const categoryMap = [
+    { value: "", display: "All Categories" },
+    { value: "Dearcare_Llp", display: "DearCare LLP" },
+    { value: "Tata_Homenursing", display: "Tata HomeNursing" }
+  ];
+
   return (
     <div className="bg-gray-50 rounded-lg border border-gray-200 overflow-hidden">
       <div className="flex flex-col sm:flex-row gap-3 sm:items-center sm:justify-between p-4 border-b border-gray-200">
@@ -72,6 +83,50 @@ export function AttendanceHeader({
               onChange={(e) => setSearchTerm(e.target.value)}
               className="pl-10 w-full bg-white text-sm text-gray-800 placeholder:text-gray-400 border-gray-200 h-10 focus-visible:ring-blue-400 focus-visible:ring-offset-2 shadow-sm rounded-lg border px-4 py-2"
             />
+          </div>
+        </div>
+        
+        {/* Category filter section */}
+        <div className="mt-2">
+          <div className="flex items-center gap-2 mb-2">
+            <span className="text-xs font-medium text-gray-600 whitespace-nowrap">Category:</span>
+            <div className="flex gap-1.5 items-center flex-wrap">
+              {categoryMap.map((category) => (
+                <button
+                  key={category.value}
+                  onClick={() => handleCategoryChange(category.value)}
+                  className={`px-2.5 py-1 rounded-md text-xs font-medium transition-colors ${
+                    selectedCategory === category.value
+                      ? "bg-blue-50 text-blue-700 border border-blue-200"
+                      : "bg-white text-gray-600 hover:bg-gray-100 border border-gray-200"
+                  }`}
+                >
+                  {category.display}
+                </button>
+              ))}
+            </div>
+          </div>
+          
+          {/* Mobile view for category filter */}
+          <div className="sm:hidden mt-2">
+            <select
+              value={selectedCategory}
+              onChange={(e) => handleCategoryChange(e.target.value)}
+              className="w-full rounded-md border border-gray-200 bg-white py-1.5 px-2 text-sm text-gray-800 appearance-none focus:outline-none focus:ring-1 focus:ring-blue-400"
+              style={{ 
+                backgroundImage: `url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' fill='none' viewBox='0 0 20 20'%3e%3cpath stroke='%236b7280' stroke-linecap='round' stroke-linejoin='round' stroke-width='1.5' d='M6 8l4 4 4-4'/%3e%3c/svg%3e")`,
+                backgroundPosition: `right 0.5rem center`,
+                backgroundRepeat: `no-repeat`,
+                backgroundSize: `1.5em 1.5em`,
+                paddingRight: `2rem`
+              }}
+            >
+              {categoryMap.map((category) => (
+                <option key={category.value} value={category.value}>
+                  {category.display}
+                </option>
+              ))}
+            </select>
           </div>
         </div>
       </div>
